requirements.txt
pandas>=2.0.0
torch>=2.0.0
tsai
requests>=2.28.0
python-dotenv>=1.0.0
